requirements.txt
APScheduler==3.11.0
black==25.1.0
blinker==1.9.0
cfgv==3.4.0
click==8.1.8
distlib==0.3.9
dnspython==2.7.0
filelock==3.18.0
Flask==3.1.0
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
identify==2.6.12
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
pip-check==3.1
platformdirs==4.3.8
pre_commit==4.2.0
PyJWT==2.10.1
pymongo==4.12.1
pyotp==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
qrcode==8.2
ruff==0.11.8
terminaltables==3.1.10
tzlocal==5.3.1
virtualenv==20.31.2
Werkzeug==3.1.3
